+++ packages/article-extras/__tests__/shared.web.js
@@ -33,16 +33,20 @@ export default () => {
     {
       name: "renders correctly",
       test: () => {
+        const isAllowed = true;
         const testInstance = TestRenderer.create(
           <ArticleExtras
             analyticsStream={() => {}}
             articleId="dummy-article-id"
+            commentsAllowed={isAllowed}
             commentsEnabled
             registerNode={() => {}}
+            relatedArticleAllowed={isAllowed}
             relatedArticleSlice={relatedArticleSlice}
             relatedArticlesVisible
             spotAccountId="dummy-spot-id"
             topics={topics}
+            topicsAllowed={isAllowed}
           />
         );
 
@@ -52,16 +56,20 @@ export default () => {
     {
       name: "no related articles, topics and comments when user not logged in",
       test: () => {
+        const isAllowed = false;
         const testInstance = TestRenderer.create(
           <ArticleExtras
             analyticsStream={() => {}}
             articleId="dummy-article-id"
+            commentsAllowed={isAllowed}
             commentsEnabled
             registerNode={() => {}}
-            relatedArticleSlice={null}
+            relatedArticleAllowed={isAllowed}
+            relatedArticleSlice={relatedArticleSlice}
             relatedArticlesVisible
-            spotAccountId={null}
-            topics={null}
+            spotAccountId="dummy-spot-id"
+            topics={topics}
+            topicsAllowed={isAllowed}
           />
         );
